### 2022-11-30 #4984 @bpm5026
I Kristen,

Sorry, I had forgotten to make the repo public. I have switched that
setting now, so let me know if that fixes the problem.

Thanks,
Brandon

On Wed, Nov 30, 2022 at 2:47 PM Kristen Thyng ***@***.***>
wrote:

> Hi @bpm5026 <https://github.com/bpm5026> and thanks for your submission.
> Did you perhaps give the wrong address for the JOSS paper? The system is
> having trouble finding it.
>
