### 2022-11-30 #4984 @kthyng
Thanks @bpm5026, though [we require](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) the paper to be in the same repository as the software:
> Your paper (paper.md and BibTeX files, plus any figures) must be hosted in a Git-based repository together with your software (although they may be in a short-lived branch which is never merged with the default).

Could you move the paper to the software repo?
