### 2022-11-30 a69ddfd
Delete paper.md
