+++ paper.bib
@@ -1,0 +1,240 @@
+@article{Hibler:1977,
+  title={A viscous sea ice law as a stochastic average of plasticity},
+  author={Hibler III, WD},
+  journal={Journal of Geophysical Research},
+  volume={82},
+  number={27},
+  pages={3932--3938},
+  year={1977},
+  publisher={Wiley Online Library}
+}
+
+@article{Hibler:1979,
+  title={A dynamic thermodynamic sea ice model},
+  author={Hibler, W DIII},
+  journal={Journal of physical oceanography},
+  volume={9},
+  number={4},
+  pages={815--846},
+  year={1979},
+  publisher={American Meteorological Society}
+}
+
+@article{Feltham:2008,
+  title={Sea ice rheology},
+  author={Feltham, Daniel L},
+  journal={Annu. Rev. Fluid Mech.},
+  volume={40},
+  pages={91--112},
+  year={2008},
+  publisher={Annual reviews}
+}
+
+@article{Coon:2007,
+  title={Arctic Ice Dynamics Joint Experiment (AIDJEX) assumptions revisited and found inadequate},
+  author={Coon, Max and Kwok, Ron and Levy, Gad and Pruis, Matthew and Schreyer, Howard and Sulsky, Deborah},
+  journal={Journal of Geophysical Research: Oceans},
+  volume={112},
+  number={C11},
+  year={2007},
+  publisher={Wiley Online Library}
+}
+
+@article{Hunke:1997,
+  title={An elastic--viscous--plastic model for sea ice dynamics},
+  author={Hunke, Elizabeth C and Dukowicz, John K},
+  journal={Journal of physical oceanography},
+  volume={27},
+  number={9},
+  pages={1849--1867},
+  year={1997}
+}
+
+@article{Rampal:2016,
+  title={neXtSIM: a new Lagrangian sea ice model},
+  author={Rampal, Pierre and Bouillon, Sylvain and {\'O}lason, Einar and Morlighem, Mathieu},
+  journal={The Cryosphere},
+  volume={10},
+  number={3},
+  pages={1055--1073},
+  year={2016},
+  publisher={Copernicus GmbH}
+}
+
+@article{Rothrock:1984,
+  title={Measuring the sea ice floe size distribution},
+  author={Rothrock, DA and Thorndike, AS},
+  journal={Journal of Geophysical Research: Oceans},
+  volume={89},
+  number={C4},
+  pages={6477--6486},
+  year={1984},
+  publisher={Wiley Online Library}
+}
+
+@article{Zhang:2015,
+  title={Sea ice floe size distribution in the marginal ice zone: Theory and numerical experiments},
+  author={Zhang, Jinlun and Schweiger, Axel and Steele, Michael and Stern, Harry},
+  journal={Journal of Geophysical Research: Oceans},
+  volume={120},
+  number={5},
+  pages={3484--3498},
+  year={2015},
+  publisher={Wiley Online Library}
+}
+
+@article{Stern:2018,
+    author = {Stern, Harry L. and Schweiger, Axel J. and Zhang, Jinlun and Steele, Michael},
+    title = "{On reconciling disparate studies of the sea-ice floe size distribution}",
+    journal = {Elementa: Science of the Anthropocene},
+    volume = {6},
+    year = {2018},
+    month = {07},
+    issn = {2325-1026},
+    doi = {10.1525/elementa.304},
+    eprint = {https://online.ucpress.edu/elementa/article-pdf/doi/10.1525/elementa.304/471884/304-5345-1-pb.pdf},
+}
+
+@article{Hutter:2020,
+  title={Feature-based comparison of sea ice deformation in lead-permitting sea ice simulations},
[…]
