### 2022-11-30 #4984 @bpm5026
Hi Kristen,

Sorry about the mixup. I have included the new branch I created off the
software with everything in it below. I just tested it, and the paper looks
to have been compiled. Let me know if sending you the branch here is okay
or if I need to resubmit with the new link.

https://github.com/SeaIce-Math/SubZero/tree/JOSS

Thanks,
Brandon

On Wed, Nov 30, 2022 at 3:08 PM Kristen Thyng ***@***.***>
wrote:

> Thanks @bpm5026 <https://github.com/bpm5026>, though we require
> <https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements>
> the paper to be in the same repository as the software:
>
> Your paper (paper.md and BibTeX files, plus any figures) must be hosted in
> a Git-based repository together with your software (although they may be in
> a short-lived branch which is never merged with the default).
>
> Could you move the paper to the software repo?
>
