### 2022-12-01 #4984 @kthyng
@bpm5026 Since you are simultaneously publishing with JAMES, could you explain in a bit of detail the distinction between what is in the other paper vs. this submission? We have to be sure to have distinct scholarly units.
