### 2022-12-01 #4984 @bpm5026
Hi Kristen,

The JAMES paper is a 30+ page paper going into the detailed model physics.
For example, in the JAMES paper, we discuss the specifics of the methods
for the contact laws, fracture rules, and things of that nature. Whereas
the intention of this JOSS paper is a short, high-level paper that we wish
to be reachable to a broader audience in the software community.

Thanks,
Brandon

On Thu, Dec 1, 2022 at 8:28 AM Kristen Thyng ***@***.***>
wrote:

> @bpm5026 <https://github.com/bpm5026> Since you are simultaneously
> publishing with JAMES, could you explain in a bit of detail the distinction
> between what is in the other paper vs. this submission? We have to be sure
> to have distinct scholarly units.
>
