### 2022-12-01 #4984 @kthyng
Thanks for that info, @bpm5026. Is the software being reviewed in the JAMES paper? And you mentioned at submission that there is a pdf on arxiv I think — could you post it here?
