### 2022-12-01 #4984 @bpm5026
Here is the link to the arxiv: https://eartharxiv.org/repository/view/3503/

The JAMES paper is not really being reviewed for the software, what they
are reviewing is model physics and how it can reproduce the statistical
behavior of the observed sea ice, such as the power-law appearance of the
floe size distribution and the long-tailed ice thickness distribution.

On Thu, Dec 1, 2022 at 12:11 PM Kristen Thyng ***@***.***>
wrote:

> Thanks for that info, @bpm5026 <https://github.com/bpm5026>. Is the
> software being reviewed in the JAMES paper? And you mentioned at submission
> that there is a pdf on arxiv I think — could you post it here?
>
