### 2022-12-06 #4984 @kthyng
@hugoledoux or @crvernon Would either of you be able to handle this submission? I know it's not a perfect fit but not sure how far it is from your normal areas of expertise.
