+++ paper.bib
@@ -202,8 +202,7 @@ pages={114–120}
 
 @article{Manucharyan:2022,
   title={SubZero: A Sea Ice Model with an Explicit Representation of the Floe Life Cycle},
-  author={Manucharyan, Georgy and 
-  , Brandon},
+  author={Manucharyan, Georgy and Montemuro, Brandon},
   year={2022},
   publisher={EarthArXiv}
 }
