### 2022-12-07 #4984 @kthyng
@hugoledoux It would be great if you could take this on.

@bpm5026 Please take a look at our list of volunteers for reviewing and list 5+ github handles (without the "@" so they aren't preemptively pinged) that would be appropriate for reviewing your software. https://bit.ly/joss-reviewers
You're also welcome to list people by github handle or email address that aren't on that list as long as you are not conflicted with them. This will help @hugoledoux get reviewers assigned to start the process more quickly. Thank you!
