### 2022-12-07 #4984 @bpm5026
Thanks Hugo! Here are 6 names from the list that say Matlab for languages,
have either oceanography or discrete element methods on their domains, and
are not listed as having any active reviews.

vsangelidakis

mvdh7

ZijieZhaoMMHW

wingtorres

chadagreene

chouj

Best,

Brandon

On Wed, Dec 7, 2022 at 6:51 AM Kristen Thyng ***@***.***>
wrote:

> @hugoledoux <https://github.com/hugoledoux> It would be great if you
> could take this on.
>
> @bpm5026 <https://github.com/bpm5026> Please take a look at our list of
> volunteers for reviewing and list 5+ github handles (without the "@" so
> they aren't preemptively pinged) that would be appropriate for reviewing
> your software. https://bit.ly/joss-reviewers
> You're also welcome to list people by github handle or email address that
> aren't on that list as long as you are not conflicted with them. This will
> help @hugoledoux <https://github.com/hugoledoux> get reviewers assigned
> to start the process more quickly. Thank you!
>
