### 2023-01-04 #4984 @ZijieZhaoMMHW
> @ZijieZhaoMMHW would you be interested in doing that review?
>
> (btw you're a hard person to email, I've tried 10min+ to find your email and I failed...)

Dear Hugo @hugoledoux ,

Thanks for the nomination of reviewing. My studies focus on meteorology and oceanography and I am not an expert in sea ice, so I should probably decline the invitation. I think @chadagreene may be a good reviewer for this as he is an expert in both MATLAB and Antarctic science.

I now add my personal email to my github profile. Sorry for the inconvenience.

Cheers,
Zijie
