### 2023-01-05 #4984 @ZijieZhaoMMHW
> Hello @ZijieZhaoMMHW, since @chadagreene cannot do it, finding a reviewer is turning out to be impossible, everyone says they are not an expert or no MATLAB expertise...
>
> JOSS is about the content but also and foremost about the software, so if you were willing to do this that would help us.

Dear Hugo @hugoledoux

Ok I can be the reviewer for this one. Another potential reviewer I know can be @ZhaohuiWang-imas, an expert in Antarctic science and MATLAB. He is also willing to do this review. However, he has no publications with JOSS, I am not sure if it can be a problem ....
