### 2023-01-05 #4984 @hugoledoux
This is super news @ZijieZhaoMMHW , thanks a lot.

@zhaohuiwang-imas Glad you can do this review. Since it's your first review for JOSS, don't hesitate to contact me (directly in the issue using @hugoledoux) if it's not clear. The review is open to everyone, and we expect you to have a dialogue with the authors, the review is often formed of several iterations.

The review will be in a new GitHub issue that I will open in a moment.
