### 2023-01-05 #5039 @editorialbot
[REVIEW]: SubZero: a Discrete element sea ice model that simulates floes as evolving concave polygons
**Submitting author:** @bpm5026 (<a href="http://orcid.org/0000-0003-1946-4916">Brandon Montemuro</a>)
**Repository:** https://github.com/SeaIce-Math/SubZero
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v1.0.4
