### 2023-02-22 #5039 @ZijieZhaoMMHW
Sorry for the delay in reviewing. My comments are provided as follows: @hugoledoux @bpm5026

Using decent parameterization and correspondences to previous literature, this versatile toolbox offers an idealized depiction of floe life cycles. To confirm its practicality, I conducted multiple tests following the instructions on the GitHub page and utilizing the three validation cases provided by the authors. As a result, I verified the efficacy of this toolbox and I think its associated paper can be published in the JOSS. However, I do have some suggestions to improve the toolbox from the view of users. My comments are listed as follows:

1.	Software environment: I have run the toolbox on two MAC systems, but I have observed that XCode is required to run the toolbox on this platform. Consequently, I am uncertain whether the program can function correctly on Windows. Can you please verify this?
2.	Validation cases: The "validation cases" file contains three comprehensive cases provided by the authors. Although these cases can be executed correctly, they can be quite challenging for novice users like myself who have limited experience with idealized floe simulations and sea ice. Therefore, I suggest that the authors provide step-by-step instructions for the three validation cases in a "readme" file, as many other authors have done in their toolboxes published in JOSS. Additionally, this paper seems to be the toolbox version of the paper “SubZero: A Sea Ice Model with an Explicit Representation of the Floe Life Cycle” published in JAMES. If you can use this toolbox to reproduce results in the source paper it can be very helpful for new users.
3.	Documents: There are some typo errors in the documents provided in the toolbox, like the “MATLBA” in the “Installation instructions”. Please correct them. Additionally, a “changing history” document file is required to record the update and changing history of the toolbox.
4.     Community guidelines: Please include a document file to indicate how others can contribute to the further development and update of this toolbox and list it in the readme file.

My proficiency in idealized floe simulations is restricted, so during the review, I focused on verifying that the toolbox functioned correctly according to the authors' description and that the resulting outputs were scientifically plausible, both of which were satisfactory. However, I am unable to assess whether the toolbox has been coded in a scientifically sound manner or if there are any inaccuracies associated with the presented floe simulations. For this matter, it would be appropriate to await comments from the other reviewer and feedback from users in the future.
