+++ README.md
@@ -70,18 +70,46 @@ To see examples of different scenarios, see the examples included in `validation
 ## Validation Cases  
 ### 1. Uniaxial Compression:
 Here we demonstrate the behavior of sea ice floes subject to uniaxial compression in a confined domain. The run is initialized with 200 floes in a fully-packed domain with the North/South boundaries moving towards the center of the domain, and stationary East/West boundaries. A relatively small time step, dt=5 s, is used to resolve the elastic waves in response to external boundary motion and changes in the floe configuration due to fractures. The atmospheric and oceanic stresses are set to zero for this simplified test. The floes are subject to Mohr-Coulomb fracture criteria, but there is no floe simplification, corner grinding, welding, ridging, rafting, or creation of new floes in this scenario. The boundaries move with a constant prescribed velocity, v_b = 0.1 m/s.
+
+1. In `Subzero.m' set:
+  1a. Flags to false except FRACTURES and COLLISION to true. 
+  1b. Set dt = 5;
+  1c. The variables `height.mean' is set to 1 and `height.delta' should be set to 0. 
+  1d. Atospheric winds should be U0 = 0 and V0 = 0. 
+  1e. Target_concentration = 1
+  1f. The number of initial floes to 200
+  1g. The modulus = 2.5e3*(mean(sqrt(cat(1,Floe).area))+min(sqrt(cat(1,Floe.area)))
+  1h. Set all instances of doInt.flag = false so there are no forces from ocean/atmosphere and all movement is driven by boundaries.
+  1i. The frequency for fractures is set to mod(i_step,200)
+  1j. After updating time and i_step add the following lines to have the boundary move.
+    if max(c2_boundary(2,:)) > 85000 && mod(i_step,30)==0
+        xb = c2_boundary(1,:);
+        yb = c2_boundary(2,:);
+        yb = yb - 15*[-1 1 1 -1];
+        c2_boundary = [xb; yb];
+        Ly = max(c2_boundary(2,:));Lx = max(c2_boundary(1,:));
+        c2_boundary_poly = polyshape(c2_boundary');
+        c2_border = scale(c2_boundary_poly,2); c2_border = subtract(c2_border, c2_boundary_poly);
+        floebound = initialize_floe_values(c2_border, height);
+      end
+  
+2. In `initialize_boundaries.m' the box is set such that Lx=Ly=10 kilometers.
+  
+3. In `fracture.m' set Sig11 = 1.5e5 and make sure all the elliptical yield curve lines are commented out or deleted.  
+
+4. In `floe_interactions.m' set mu = 0.3.
   
 ### 2. Nares Strait:
[…]
