+++ README.md
@@ -69,7 +69,7 @@ To see examples of different scenarios, see the examples included in `validation
 
 ## Validation Cases  
 ### 1. Uniaxial Compression:
-Here we demonstrate the behavior of sea ice floes subject to uniaxial compression in a confined domain. The run is initialized with 200 floes in a fully-packed domain with the North/South boundaries moving towards the center of the domain, and stationary East/West boundaries. A relatively small time step, dt=5 s, is used to resolve the elastic waves in response to external boundary motion and changes in the floe configuration due to fractures. The atmospheric and oceanic stresses are set to zero for this simplified test. The floes are subject to Mohr-Coulomb fracture criteria, but there is no floe simplification, corner grinding, welding, ridging, rafting, or creation of new floes in this scenario. The boundaries move with a constant prescribed velocity, v_b = 0.1 m/s.
+Here we demonstrate the behavior of sea ice floes subject to uniaxial compression in a confined domain. The run is initialized with 200 floes in a fully-packed domain with the North/South boundaries moving towards the center of the domain, and stationary East/West boundaries. A relatively small time step, dt=5 s, is used to resolve the elastic waves in response to external boundary motion and changes in the floe configuration due to fractures. The atmospheric and oceanic stresses are set to zero for this simplified test. The floes are subject to Mohr-Coulomb fracture criteria, but there is no floe simplification, corner grinding, welding, ridging, rafting, or creation of new floes in this scenario. The boundaries move with a constant prescribed velocity, v_b = 0.1 m/s. To modify the base program to run the uniaxial compression validation case on your own, the following changes must be made.
 
 1. In ```Subzero.m``` set:
   
@@ -104,14 +104,16 @@ Here we demonstrate the behavior of sea ice floes subject to uniaxial compressio
             floebound = initialize_floe_values(c2_border, height);  
         end
   
-2. In ```initialize_boundaries.m``` the box is set such that Lx=Ly=10 kilometers.
+2. In ```initialize_boundaries.m``` the box is set such that Lx=Ly=100 kilometers.
   
 3. In ```fracture.m``` set Sig11 = 1.5e5 and make sure all the elliptical yield curve lines are commented out or deleted.  
 
 4. In ```floe_interactions.m``` set mu = 0.3.
   
+5. To do the other cases, rerun with modulus using 2.0e3 and 1.5e3 in place of 2.5e3.
+  
 ### 2. Nares Strait:
-The Nares Strait simulation demonstrates the role of floe fractures in wind-driven sea ice transport through narrow straits. Nares Strait is a channel between Ellesmere Island (Canada) and Greenland. The simulation aims to reflect spring or summer-like conditions of Arctic sea ice export through Nares Strait after the breakup of its winter arches. Since the transport events are relatively short (order of days or less), the effects of thermodynamic sea ice melt could be considered secondary relative to mechanical floe processes such as collisions and fractures. We thus randomly initialize the model with relatively large floes of uniform thickness, covering only the area just north of the strait. The uniform 10 m/s southward winds generate stresses that push the floes through the strait, while the ocean is assumed to be stagnant. Coastal boundaries are prescribed using a series of static floes. All physical processes except collisions and fractures are turned off to model the spring/summer breakup of floes. To modify the program to run the Nares validation case on your own, the following changes must be made.
[…]
