### 2023-03-30 #5039 @bpm5026
@hugoledoux Sorry for the delay, I have addressed the majority of the review comments. I just underestimated the amount of time it would take to write the automated test. I have most of the automated test written and will be finished once that is done. I think this should be finished relatively soon (hopefully by the end of the weekend) and will let you know once I am done.
