+++ CHANGES.md
@@ -1,3 +1,9 @@
+1.0.2 (April 3, 2023)
+==================
+
+- Updates for JOSS paper, including updates to README.md, fixing a bug in calculating collisions between ghost floes, the addition of an automated test, and additional documentation. New documentation includes the CHANGES.md file and CONTRIBUTING.md.
+
+
 1.0.2 (Nov 18, 2022)
 ==================
 
