### 2023-04-03 #5039 @bpm5026
@hugoledoux  @ZijieZhaoMMHW Thank you again for the thoughtful reviews; here is what has been addressed. Let me know if there are any additional comments or if you feel I have not adequately addressed the original comments.
1) When we tested the program to function correctly on Windows, it ran fine without any additional installations similar to Xcode on the mac.
2) Step-by-step instructions have been added to the README file for modifying the base code to run the validation cases.
3) Updated typographic errors in the readme and added CHANGES.md file to the repository.
4) Added CONTRIBUTING.md file with instructions.
5) Added automated test, which tests to ensure that collisions do not generate energy.
