### 2023-06-07 14e5b9c
Update README.md
Updated some bugs in changes describing how to set up Nares Strait validation case
