### 2023-06-08 23e0519
Update README.md
added Nb = ...
