+++ README.md
@@ -164,6 +164,8 @@ The Nares Strait simulation demonstrates the role of floe fractures in wind-driv
         end
         Bu = union([Floe.poly]);
         Bu = union(Bu,B);
+    
+        Nb = length(Floe);
 
     4b. Then, after ```[~, b,~,~,~] = polybnd_voronoi([X Y],boundary.Vertices)``` insert  the following lines to remove any overlap between the shapes generated by the Voronoi tesselation and Nares Strait topography.
 
