### 2023-06-10 9ab78b8
Update create_new_ice.m
remove extemporaneous && dhdt > 0
