### 2023-06-10 0d28587
Update calc_trajectory.m
switch / to * which it should have been in frac calculations
