### 2023-06-10 02e1527
Update calc_trajectory.m
switch / to * which it should have been in frac calculations
