### 2023-07-20 #5039 @ZhaohuiWang-imas
Review of "SubZero: a Discrete element sea ice model that simulates floes as evolving concave polygons"

This article introduces a new discrete element sea ice model that utilises parameterisations of floe-scale processes. Since the theoretical manuscript of the model have already been published in JAMES and undergone peer review, this revision focused primarily on the feasibility of the model described in the article. Compared to traditional sea ice models commonly written in Fortran, this model runs on Matlab, making it easier to understand, use, and modify. It greatly simplifies the entry barrier for beginners. The program effectively distinguishes between the user-defined areas and the code areas that users do not need to touch. Upon testing, the program is plug-and-play and runs smoothly. The base program runs at a reasonable speed (tested on a 2021 MacBook Pro M1 Max chip). The three validation cases provided by the author can all be executed, and the results are as expected. The application guide (README) for the program is written in a clear and understandable manner, with a logical structure. I believe this model is very innovative, and the research findings are at the cutting edge. The author's open-access approach to the software and the clear user guide make a significant contribution to the field. Therefore, I recommend that your journal publish this important achievement.

Results: Accept for publication without revision.
