### 2023-07-28 #5039 @hugoledoux
Thanks everyone, it took some time but we got there in the end :)

@ZhaohuiWang-imas: you checked all but one check (State of the field) but I believe this is a mistake? The paper clearly has that part written, so I'll move on to accept it.

Before we continue though, 4 things need to be done:
  1. accept the PR I just made: https://github.com/SeaIce-Math/SubZero/pull/5
  2. Fix the DOIs! At JOSS we like DOIs and we want all papers to have a DOI, so please follow the tips from our bot and/or find the DOIs and add them to the bibtex file.
  3. one potential typo in the paper, line 48: shouldn't it be "CONVEX polygonal elements"? I mean, the Voronoi are convex, and after the first collisions the polygons all look convex to me
  4. line 19 + line 22: I am a layman in your field, but why the "O" for the scale? Is that a normal nomenclature? I found that confusing, but perhaps it's me 🤷‍♂️
