### 2023-07-28 398acd8
Update paper.bib
Added doi to some of the references
