+++ paper.bib
@@ -6,7 +6,8 @@
   number={27},
   pages={3932--3938},
   year={1977},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {https://doi.org/10.1029/JC082i027p03932}
 }
 
 @article{Hibler:1979,
@@ -17,7 +18,8 @@
   number={4},
   pages={815--846},
   year={1979},
-  publisher={American Meteorological Society}
+  publisher={American Meteorological Society},
+  doi = {https://doi.org/10.1175/1520-0485(1979)009<0815:ADTSIM>2.0.CO;2}
 }
 
 @article{Feltham:2008,
@@ -27,7 +29,8 @@
   volume={40},
   pages={91--112},
   year={2008},
-  publisher={Annual reviews}
+  publisher={Annual reviews},
+  doi = {https://doi.org/10.1146/annurev.fluid.40.111406.102151}
 }
 
 @article{Coon:2007,
@@ -37,7 +40,8 @@
   volume={112},
   number={C11},
   year={2007},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {https://doi.org/10.1029/2005JC003393}
 }
 
 @article{Hunke:1997,
@@ -47,7 +51,8 @@
   volume={27},
   number={9},
   pages={1849--1867},
-  year={1997}
+  year={1997},
+  doi = {https://doi.org/10.1175/1520-0485(1997)027<1849:AEVPMF>2.0.CO;2}
 }
 
 @article{Rampal:2016,
@@ -58,7 +63,8 @@
   number={3},
   pages={1055--1073},
   year={2016},
-  publisher={Copernicus GmbH}
+  publisher={Copernicus GmbH},
+  doi = {https://doi.org/10.5194/tc-10-1055-2016}
 }
 
 @article{Rothrock:1984,
@@ -69,7 +75,8 @@
   number={C4},
   pages={6477--6486},
   year={1984},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {https://doi.org/10.1029/JC089iC04p06477}
 }
 
 @article{Zhang:2015,
@@ -80,7 +87,8 @@
   number={5},
   pages={3484--3498},
   year={2015},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {https://doi.org/10.1002/2015JC010770}
 }
 
 @article{Stern:2018,
@@ -103,7 +111,8 @@
   number={1},
   pages={93--113},
   year={2020},
-  publisher={Copernicus GmbH}
+  publisher={Copernicus GmbH},
+  doi = {https://doi.org/10.5194/tc-14-93-2020}
 }
 
 @article{Mohammadi:2020,
@@ -113,7 +122,8 @@
   volume={148},
   number={8},
   pages={3287--3303},
-  year={2020}
+  year={2020},
+  doi = {https://doi.org/10.1175/MWR-D-19-0359.1}
 }
 
 @article{Hutter:2022,
@@ -124,7 +134,8 @@
   number={4},
   pages={e2021JC017666},
   year={2022},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {https://doi.org/10.1029/2021JC017667.}
 }
 
 @article{Mehlmann:2021,
@@ -135,7 +146,8 @@
   number={11},
   pages={e2021MS002523},
   year={2021},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {https://doi.org/10.1029/2021MS002523}
 }
 
 @article{Cundall:1979,
@@ -146,7 +158,8 @@
   number={1},
   pages={47--65},
   year={1979},
-  publisher={Thomas Telford Ltd}
[…]
