### 2023-07-28 0473dfe
Update paper.bib
Fixed typo in MATLAB reference
