+++ paper.bib
@@ -234,8 +234,7 @@ year = {2020},
 author = {MATLAB},
 title = {version 9.8.0 (R2020a)},
 publisher = {The MathWorks Inc.},
-address = {Natick, Massachusetts},
-doi = {
+address = {Natick, Massachusetts}
 }
 
 @software{Montemuro:2022,
