### 2023-07-28 #5039 @bpm5026
@hugoledoux  Thank you for your help, time and effort. See responses below to your comments:
1) Done
2) Added DOI to all the references that have them.
3) One of the new things about this model is we allow for concave polygons; you can see in fig 1a that the floes near the coastline are indeed concave. Then it is true that in this simulation, as they evolve, they tend to become convex, but that doesn't happen in others.
4) The big 'O' is a notation that indicates something is "on the order of," so in this case it means the floes are order 100 kilometers in size. If you would prefer that be spelled out more for this journal, let me know and I will change that.
