### 2023-07-31 #5039 @hugoledoux
@bpm5026

2. you need to remove the 'http://doi.org/' prefixes, see above
3. okay clear, maybe a sentence about this in the paper would help the reader? Not a necessity, but I think it would be nice.
4. Never seen the big-oh in that context (more O(n^2)), but it's fine, was just checking
