### 2023-07-31 8d7c375
Update paper.bib
Fixing DOI links
