+++ paper.bib
@@ -7,7 +7,7 @@
   pages={3932--3938},
   year={1977},
   publisher={Wiley Online Library},
-  doi = {https://doi.org/10.1029/JC082i027p03932}
+  doi = {0.1029/JC082i027p03932}
 }
 
 @article{Hibler:1979,
@@ -19,7 +19,7 @@
   pages={815--846},
   year={1979},
   publisher={American Meteorological Society},
-  doi = {https://doi.org/10.1175/1520-0485(1979)009<0815:ADTSIM>2.0.CO;2}
+  doi = {10.1175/1520-0485(1979)009<0815:ADTSIM>2.0.CO;2}
 }
 
 @article{Feltham:2008,
@@ -30,7 +30,7 @@
   pages={91--112},
   year={2008},
   publisher={Annual reviews},
-  doi = {https://doi.org/10.1146/annurev.fluid.40.111406.102151}
+  doi = {10.1146/annurev.fluid.40.111406.102151}
 }
 
 @article{Coon:2007,
@@ -41,7 +41,7 @@
   number={C11},
   year={2007},
   publisher={Wiley Online Library},
-  doi = {https://doi.org/10.1029/2005JC003393}
+  doi = {10.1029/2005JC003393}
 }
 
 @article{Hunke:1997,
@@ -52,7 +52,7 @@
   number={9},
   pages={1849--1867},
   year={1997},
-  doi = {https://doi.org/10.1175/1520-0485(1997)027<1849:AEVPMF>2.0.CO;2}
+  doi = {10.1175/1520-0485(1997)027<1849:AEVPMF>2.0.CO;2}
 }
 
 @article{Rampal:2016,
@@ -64,7 +64,7 @@
   pages={1055--1073},
   year={2016},
   publisher={Copernicus GmbH},
-  doi = {https://doi.org/10.5194/tc-10-1055-2016}
+  doi = {10.5194/tc-10-1055-2016}
 }
 
 @article{Rothrock:1984,
@@ -76,7 +76,7 @@
   pages={6477--6486},
   year={1984},
   publisher={Wiley Online Library},
-  doi = {https://doi.org/10.1029/JC089iC04p06477}
+  doi = {10.1029/JC089iC04p06477}
 }
 
 @article{Zhang:2015,
@@ -88,7 +88,7 @@
   pages={3484--3498},
   year={2015},
   publisher={Wiley Online Library},
-  doi = {https://doi.org/10.1002/2015JC010770}
+  doi = {10.1002/2015JC010770}
 }
 
 @article{Stern:2018,
@@ -112,7 +112,7 @@
   pages={93--113},
   year={2020},
   publisher={Copernicus GmbH},
-  doi = {https://doi.org/10.5194/tc-14-93-2020}
+  doi = {10.5194/tc-14-93-2020}
 }
 
 @article{Mohammadi:2020,
@@ -123,7 +123,7 @@
   number={8},
   pages={3287--3303},
   year={2020},
-  doi = {https://doi.org/10.1175/MWR-D-19-0359.1}
+  doi = {10.1175/MWR-D-19-0359.1}
 }
 
 @article{Hutter:2022,
@@ -135,7 +135,7 @@
   pages={e2021JC017666},
   year={2022},
   publisher={Wiley Online Library},
-  doi = {https://doi.org/10.1029/2021JC017667.}
+  doi = {10.1029/2021JC017667.}
 }
 
 @article{Mehlmann:2021,
@@ -147,7 +147,7 @@
   pages={e2021MS002523},
   year={2021},
   publisher={Wiley Online Library},
-  doi = {https://doi.org/10.1029/2021MS002523}
+  doi = {10.1029/2021MS002523}
 }
 
[…]
