### 2023-07-31 ac521be
Update paper.md
Added clarification on the O() notation and concave nature of the floes
