+++ paper.md
@@ -30,17 +30,17 @@ bibliography: paper.bib
 
 SubZero is a conceptually new discrete element sea ice model geared to explicitly simulate the life cycles of individual floes by using polygonal elements with time-evolving boundaries. This unique model uses parameterizations of floe-scale processes, such as collisions, rafting, ridging, fracturing, and welding, to simulate the behavior on sea ice floes subject to mechanical and thermodynamic forcing in confined or periodic domains. SubZero enables the exploration of a wide range of floe interaction rules and fracture criteria to further our understanding of sea ice mechanics, including distributions of floe sizes, thicknesses, and shapes. \autoref{fig:Nares} shows snapshots from a validation study where SubZero simulates floes moving through a channel. A complete model description and example process studies demonstrating its capabilities can be found in @Manucharyan:2022. SubZero was developed using MATLAB [@MATLAB:2020], leveraging its built-in functions for polygonal operations. The model source code has been archived to Zenodo [@Montemuro:2022]. 
 
-![The evolution of sea ice floes as they pass through the Nares Strait (between Canada and Greenland), including (a) initial floe state with the inset showing the location of Nares Strait, (b) floes shortly after sea ice breakup that occurred after about three days, and (c) floe state after ten days when many floes have passed through the Nares Strait. The initial distribution of floes was generated using a Voronoi tessellation, and the subsequent evolution of floe shapes is only subject to floe fractures. The green box in panel (a) shows the modeled part of Nares Strait. The blue arrows represent sea ice velocity after averaging floe momentum on an Eulerian grid.\label{fig:Nares}](Nares_Floes.png)
+![The evolution of sea ice floes as they pass through the Nares Strait (between Canada and Greenland), including (a) initial floe state with the inset showing the location of Nares Strait, (b) floes shortly after sea ice breakup that occurred after about three days, and (c) floe state after ten days when many floes have passed through the Nares Strait. The initial distribution of floes was generated using a Voronoi tessellation, and the subsequent evolution of floe shapes is only subject to floe fractures. In this simulation, the complex initial floe shapes near the land become simpler and more convex as they break up. The green box in panel (a) shows the modeled part of Nares Strait. The blue arrows represent sea ice velocity after averaging floe momentum on an Eulerian grid.\label{fig:Nares}](Nares_Floes.png)
 
 
 
 # Statement of need
 
-Sea ice dynamics span a wide range of scales. At length scales O(10--100) km and smaller, sea ice exhibits granular behavior as individual floes and fracture networks become evident [@Rothrock:1984; @Zhang:2015; @Stern:2018]. Sea ice motion at relatively large scales, O(100 km), is commonly represented in climate models using continuous rheological models, like the viscous-plastic model [@Hibler:1979], which have not been formally derived from basic sea ice physics but have been postulated instead. The rheology defines a relationship between sea ice stress caused by floe-floe or floe-lead interactions, to the large scale deformation of ice cover, the material properties of sea ice, and the state of the ice cover. As such, the rheological models carry potentially large uncertainties in representing sea ice dynamics. They are also not designed to represent the scales of motion at which individual floes start to influence dynamics [@Coon:2007]. 
[…]
