### 2023-07-31 #5039 @bpm5026
@hugoledoux

2. Removed the 'http://doi.org/' prefixes
3. Added 'In this simulation, the complex initial floe shapes near the land become simpler and more convex as they break up.' to the figure caption as well as the sentence 'Floes undergo many processes that affect their shapes, including fracturing, ridging/rafting, and welding, making them concave.' to the last paragraph before the mention of concave there.
4. Tweaked sentence so it says 'At length scales O(10--100) km and smaller (where O designates the order of magnitude), sea ice exhibits granular behavior as individual floes and fracture networks become evident' to be clear what we mean here to readers.
