### 2023-08-01 #5039 @hugoledoux
What I should have said is: modify the DOIs and run `@editorialbot check references`, there are two small errors still, see above. 1) it should start with "1" (so "10.1029..."); 2) remove the trailing dot

At this point could you:
- [ ] fix the 2 DOIs + ensure all is good
