### 2023-08-01 eac28e1
Update paper.bib
Fixing broken DOIs
