+++ paper.bib
@@ -7,7 +7,7 @@
   pages={3932--3938},
   year={1977},
   publisher={Wiley Online Library},
-  doi = {0.1029/JC082i027p03932}
+  doi = {10.1029/JC082i027p03932}
 }
 
 @article{Hibler:1979,
@@ -135,7 +135,7 @@
   pages={e2021JC017666},
   year={2022},
   publisher={Wiley Online Library},
-  doi = {10.1029/2021JC017667.}
+  doi = {10.1029/2021JC017667}
 }
 
 @article{Mehlmann:2021,
