### 2023-08-01 #5039 @bpm5026
1) Fixed the 2 DOIs (See above)
2) The release with JOSS is SubZero v1.0.4: https://github.com/SeaIce-Math/SubZero/releases/tag/v1.0.4
3) Zenodo archive: https://zenodo.org/record/8205778
4) Zenodo title matches paper, authors are listed, and author ORCID IDs are correct
5) Zenodo DOI: 10.5281/zenodo.8205778

Let me know if there is anything missing
