### 2023-08-03 #5039 @hugoledoux
> 4\. Zenodo title matches paper, authors are listed, and author ORCID IDs are correct

no they don't, please fix:

1. remove the capital "D" from "Discrete" the title in zenodo
2. the middle name of both authors need to be there in zenodo (or removed from the paper since your ORCID doesn't have them, your choice).
