### 2023-08-03 #5039 @bpm5026
Sorry, I missed those little details. I changed it to "discrete" and added the middle initials to Zenodo.
