### 2023-08-07 #5039 @hugoledoux
@kthyng I assumed it's an oversight from him, I wrote it there: https://github.com/openjournals/joss-reviews/issues/5039#issuecomment-1655103062

so all is good 👍
